apply suggestion
